user controller testing is done

+++ backend/src/utils/cloudinary.js
@@ -2,23 +2,31 @@ import { v2 as cloudinary } from "cloudinary"
 import fs from "fs"
 
 cloudinary.config({
-    account_id: process.env.CLOUDINARY_CLOUD_NAME,
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
-    api_secre: process.env.CLOUDINARY_API_SECRET,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
 })
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null;
 
+        console.log(`Uploading file to Cloudinary: ${localFilePath}`);
+        
+
         const res = await cloudinary.uploader.upload(localFilePath)
 
-        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath)
+        }
         return res
     } catch (error) {
         console.log(`something want wrong while upload the image ${error}`);
-        fs.unlinkSync(localFilePath)
-        return error
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath)
+        }
+        throw new Error("Cloudinary upload failed: " + error.message);
     }
 }
